move around commands and command perms
