+++ src/main/java/client/command/commands/gm0/ResetStatsCommand.java
@@ -26,24 +26,20 @@ package client.command.commands.gm0;
 import client.Character;
 import client.Client;
 import client.command.Command;
-import net.server.Server;
-import net.server.channel.Channel;
 
-public class OnlineCommand extends Command {
+public class ResetStatsCommand extends Command {
     {
-        setDescription("Show all online players.");
+        setDescription("Set all primary stats to 4 and returns ap.");
     }
 
     @Override
     public void execute(Client c, String[] params) {
         Character player = c.getPlayer();
-        for (Channel ch : Server.getInstance().getChannelsFromWorld(player.getWorld())) {
-            player.yellowMessage("Players in Channel " + ch.getId() + ":");
-            for (Character chr : ch.getPlayerStorage().getAllCharacters()) {
-                if (!chr.isGM()) {
-                    player.message(" >> " + Character.makeMapleReadable(chr.getName()) + " is at " + chr.getMap().getMapName() + ".");
-                }
-            }
+
+        try {
+            int spUsed = player.getStr()-4 + player.getDex()-4 + player.getInt()-4 + player.getLuk()-4;
+            player.updateStrDexIntLuk(4, spUsed);
+        } catch (NumberFormatException nfe) {
         }
     }
 }
